+++ pkg/machine/define/vmtype.go
@@ -12,6 +12,7 @@ const (
 	WSLVirt
 	AppleHvVirt
 	HyperVVirt
+	LibKrun
 	UnknownVirt
 )
 
@@ -22,6 +23,7 @@ const (
 	qemu    = "qemu"
 	appleHV = "applehv"
 	hyperV  = "hyperv"
+	libkrun = "libkrun"
 )
 
 func (v VMType) String() string {
@@ -29,6 +31,23 @@ func (v VMType) String() string {
 	case WSLVirt:
 		return wsl
 	case AppleHvVirt:
+		return appleHV
+	case HyperVVirt:
+		return hyperV
+	case LibKrun:
+		return libkrun
+	}
+	return qemu
+}
+
+// DiskType returns a string representation that matches the OCI artifact
+// type on the container image registry
+func (v VMType) DiskType() string {
+	switch v {
+	case WSLVirt:
+		return wsl
+		// Both AppleHV and Libkrun use same raw disk flavor
+	case AppleHvVirt, LibKrun:
 		return appleHV
 	case HyperVVirt:
 		return hyperV
@@ -44,6 +63,8 @@ func (v VMType) ImageFormat() ImageFormat {
 		return Raw
 	case HyperVVirt:
 		return Vhdx
+	case LibKrun:
+		return Raw
 	}
 	return Qcow
 }
@@ -56,6 +77,8 @@ func ParseVMType(input string, emptyFallback VMType) (VMType, error) {
 		return WSLVirt, nil
 	case appleHV:
 		return AppleHvVirt, nil
+	case libkrun:
+		return LibKrun, nil
 	case hyperV:
 		return HyperVVirt, nil
 	case "":
