+++ pkg/machine/vmconfigs/config_darwin.go
@@ -3,12 +3,16 @@ package vmconfigs
 import (
 	"os"
 
-	"github.com/containers/podman/v5/pkg/machine/applehv/vfkit"
+	"github.com/containers/podman/v5/pkg/machine/apple/vfkit"
 )
 
 type AppleHVConfig struct {
 	// The VFKit endpoint where we can interact with the VM
-	Vfkit vfkit.VfkitHelper
+	Vfkit vfkit.Helper
+}
+
+type LibKrunConfig struct {
+	KRun vfkit.Helper
 }
 
 // Stubs
